requirements.txt
Pillow>=9.0.0
py2app>=0.28